requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorzero==2.0
distlib==0.3.9
filelock==3.17.0
gpiozero==2.0.1
idna==3.10
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.1
